requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
neo4j==5.14.1
sse-starlette==1.8.2
pydantic==2.5.0
python-multipart==0.0.6
sseclient-py==1.8.0
requests==2.31.0 